Reconcile state save and switch to wss
